playing around with bash

+++ lib/utils.sh
@@ -7,5 +7,19 @@ function create_target_folder {
 }
 
 function clean_folder {
-  rm -r $1
+  if [ -d $1 ]
+  then
+    rm -r $1
+  else
+    echo "could not clean folder $1, due to folder doesn't exist"
+  fi
 }
+
+function copy_src {
+  cp -r "$1/." "$2"
+}
+
+function find_test_functions {
+  echo $(grep -A1 '#@test' $1 | grep 'function .* {' | cut -d ' ' -f 2)
+}
+
